feat(goal): transaction 지출 내역 goal과 연동

+++ backend/armageddon/src/main/java/com/aespa/armageddon/core/domain/goal/repository/GoalRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface GoalRepository extends JpaRepository<Goal,Integer> {
+public interface GoalRepository extends JpaRepository<Goal, Long> {
     List<Goal> findByUserIdAndStatusNot(Long userId, GoalStatus status);
 
     Optional<Goal> findByGoalIdAndUserId(Long goalId, Long userId);
